+++ src/main/java/org/cip4/lib/xjdf/exception/XJdfInitException.java
@@ -1,0 +1,12 @@
+package org.cip4.lib.xjdf.exception;
+
+public class XJdfInitException extends Exception {
+
+    public XJdfInitException(Throwable throwable) {
+        super(throwable);
+    }
+
+    public XJdfInitException(String message, Throwable throwable) {
+        super(message, throwable);
+    }
+}
