+++ src/main/java/org/cip4/lib/xjdf/util/XJdfLib.java
@@ -1,6 +1,7 @@
 package org.cip4.lib.xjdf.util;
 
 import jakarta.xml.bind.JAXBException;
+import org.cip4.lib.xjdf.exception.XJdfInitException;
 import org.cip4.lib.xjdf.xml.internal.JAXBContextFactory;
 
 public class XJdfLib {
@@ -8,11 +9,7 @@ public class XJdfLib {
     /**
      * Helper method to initialize the library's JAXBContext.
      */
-    public static void initLibrary() throws Exception {
-        try {
-            JAXBContextFactory.init();
-        } catch (JAXBException e) {
-            throw new Exception(e);
-        }
+    public static void init() throws XJdfInitException {
+        JAXBContextFactory.init();
     }
 }
