+++ src/test/java/org/cip4/lib/xjdf/schema/UnreferencedNodesTest.java
@@ -1,13 +1,12 @@
 package org.cip4.lib.xjdf.schema;
 
 import org.cip4.lib.xjdf.xml.XJdfConstants;
-import org.cip4.lib.xjdf.xml.internal.NamespaceManager;
+import org.cip4.lib.xjdf.binding.NamespaceManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
-import org.xml.sax.InputSource;
 
 import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilder;
