+++ src/test/java/org/cip4/lib/xjdf/schema/XsdReader.java
@@ -1,7 +1,7 @@
 package org.cip4.lib.xjdf.schema;
 
 import org.cip4.lib.xjdf.xml.XJdfConstants;
-import org.cip4.lib.xjdf.xml.internal.NamespaceManager;
+import org.cip4.lib.xjdf.binding.NamespaceManager;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
