+++ src/test/java/org/cip4/lib/xjdf/xml/XJmfParser2Test.java
@@ -4,6 +4,7 @@ import org.cip4.lib.xjdf.exception.XJdfParseException;
 import org.cip4.lib.xjdf.schema.Header;
 import org.cip4.lib.xjdf.schema.XJMF;
 import org.cip4.lib.xjdf.type.DateTime;
+import org.cip4.lib.xjdf.xml.XJdfParser;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
 
@@ -14,50 +15,7 @@ import static org.junit.jupiter.api.Assertions.*;
 /**
  * JUnit test case for XJmfParser.
  */
-public class XJmfParserTest {
-
-    @Test
-    public void parseXJmf_ByteArray() throws Exception {
-        // arrange
-        final XJMF xjmf = new XJMF().withHeader(
-            new Header().withDeviceID("MyDevice").withID("ID").withTime(new DateTime("2017-03-27T21:31:54Z"))
-        );
-
-        // act
-        // assert
-        assertThrows(
-            XJdfParseException.class,
-            new Executable() {
-                @Override
-                public void execute() throws Throwable {
-                    new XJmfParser().parseXJmf(xjmf);
-                }
-            }
-        );
-    }
-
-    @Test
-    public void parseXJmf_OutputStream() throws Exception {
-        // arrange
-        final XJMF xjmf = new XJMF().withHeader(
-            new Header().withDeviceID("MyDevice").withID("ID").withTime(new DateTime("2017-03-27T21:31:54Z"))
-        );
-
-        // act
-        final ByteArrayOutputStream bos = new ByteArrayOutputStream();
-
-
-        // assert
-        assertThrows(
-            XJdfParseException.class,
-            new Executable() {
-                @Override
-                public void execute() throws Throwable {
-                    new XJmfParser().parseXJmf(xjmf, bos);
-                }
-            }
-        );
-    }
+public class XJmfParser2Test {
 
     @Test
     public void parseXJmf_ByteArray_SkipValidation() throws Exception {
@@ -67,7 +25,7 @@ public class XJmfParserTest {
         );
 
         // act
-        byte[] actual = new XJmfParser().parseXJmf(xjmf, true);
+        byte[] actual = new XJdfParser<XJMF>().writeXml(xjmf);
 
         // assert
         String jmf = new String(actual);
@@ -85,7 +43,7 @@ public class XJmfParserTest {
 
         // act
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
-        new XJmfParser().parseXJmf(xjmf, bos,true);
+        new XJdfParser<XJMF>().writeXml(xjmf, bos);
         byte[] actual = bos.toByteArray();
 
         // assert
