+++ src/test/java/org/cip4/lib/xjdf/xml/internal/JAXBContextFactoryTest.java
@@ -2,6 +2,7 @@ package org.cip4.lib.xjdf.xml.internal;
 
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBException;
+import org.cip4.lib.xjdf.exception.XJdfInitException;
 import org.junit.jupiter.api.Test;
 
 
@@ -10,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.*;
 public class JAXBContextFactoryTest {
 
     @Test
-    public void initIsIdempotent() throws JAXBException {
+    public void initIsIdempotent() throws JAXBException, XJdfInitException {
         JAXBContextFactory.init();
         JAXBContext context1 = JAXBContextFactory.getInstance();
         JAXBContextFactory.init();
