+++ src/test/java/org/cip4/lib/xjdf/xml/legacy/XJdfUnpackagerTest.java
@@ -1,4 +1,4 @@
-package org.cip4.lib.xjdf.xml;
+package org.cip4.lib.xjdf.xml.legacy;
 
 import java.io.File;
 import java.io.IOException;
