+++ src/test/resources/org/cip4/lib/xjdf/xml/validator-invalid_semantic-1.xjdf
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?><!-- Generated by CIP4 xJdfLib 0.5-SNAPSHOT -->
+<xjdf:XJDF xmlns:xjdf="http://www.CIP4.org/JDFSchema_2_0" CommentURL="./layout.jdf" ID="XJDF_PSQ131S2" JobID="MyJobId"
+           Version="2.0">
+    <xjdf:ProductList>
+        <xjdf:Product Amount="1000"/>
+    </xjdf:ProductList>
+    <xjdf:ParameterSet Name="RunList">
+        <xjdf:Parameter>
+            <xjdf:RunList>
+                <xjdf:FileSpec URL="./test.pdf"/>
+            </xjdf:RunList>
+        </xjdf:Parameter>
+        <xjdf:Parameter>
+            <xjdf:RunList>
+                <xjdf:FileSpec URL="subfolder/test2.pdf"/>
+            </xjdf:RunList>
+        </xjdf:Parameter>
+    </xjdf:ParameterSet>
+</xjdf:XJDF>
