+++ src/test/resources/org/cip4/lib/xjdf/xml/validator-invalid_syntax-1.xjdf
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<XJDF>
+<test>
+</XJDF>
