+++ src/test/resources/org/cip4/lib/xjdf/xml/validator-valid.xjdf
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<XJDF xmlns="http://www.CIP4.org/JDFSchema_2_0" Version="2.1" JobID="ab" Types="Product">
+	<Comment>aÄoÖuÜsß</Comment>
+</XJDF>
